FEAT: reorganized file structure
