+++ src/granad/_core.py
@@ -1,835 +1,10 @@
-from functools import reduce
-from typing import Callable, Union
-
-import diffrax
 import jax
 import jax.numpy as jnp
-import matplotlib.pyplot as plt
-import numpy as np
-from flax import struct
-from matplotlib.colors import Normalize
-
 jax.config.update("jax_enable_x64", True)
+import diffrax
 
-from collections.abc import Iterable
-from enum import Enum, auto
-from functools import wraps
-from typing import NewType
-
-## TYPE ALIASES
-FieldFunc = Callable[[float], jax.Array]
-DissipationFunc = Callable[[jax.Array], jax.Array]
-InputDict = NewType("InputDict", dict[str, Callable[[float], complex]])
-TransitionsDict = NewType(
-    "TransitionsDict", Union[None, dict[tuple[str, str], list[float]]]
-)
-
-# TODO: improve and update docstrings
-
-
-## CLASSES
-@struct.dataclass
-class Orbital:
-    """An orbital.
-    It is only used briefly during stack construction and immediately discarded after.
-
-     - `orbital_id`: Orbital ID, e.g. "pz_graphene".
-     - `position`: Orbital position like $(x,y,z)$, in Ångström.
-     - `occupation`: number of electrons in the non-interacting orbital (`1` or `0`)
-    """
-
-    orbital_id: str
-    position: jax.Array
-    occupation: int = 1
-    sublattice_id: int = 0
-
-
-@struct.dataclass
-class Stack:
-    """A stack of orbitals.
-
-    - `hamiltonian`: TB-Hamiltonian of the system.
-    - `coulomb`: Coulomb matrix of the system.
-    - `rho_0`: Density matrix corresponding to the initial state. If a simulation is interrupted at time t, then it can be resumed using the density matrix at time t as rho_0.
-    - `rho_stat`: Density matrix corresponding to the final state.
-    - `energies`: Independent particle energies as given by the eigenvalues of the TB-hamiltonian.
-    - `eigenvectors`: Eigenvectors of the TB-Hamiltonian. These are used to transform between *site* and *energy* basis.
-    - `positions`: Positions of the orbitals, $N \\times 3$ - dimensional.
-    - `unique_ids`: Collection of unique strings. Each string identifies an orbital type occuring in the stack. Typically something like `"pz_graphene"`.
-    - `ids`: The integer at the n-th position corresponds to the type of the n-th orbital via its index in `unique_ids`.
-    - `eps`: numerical precision threshold for identifiying degeneracies in the eigenenergies.
-    - `homo`: index of the homo, such that stack.energies[homo] is the energy of the ground state homo.
-    - `electrons`: total number of electrons in the system.
-    - `from_state`: the state from which electrons have been taken for the initial excited transition, the state translates to HOMO - from_state
-    - `to_state`: the state into which electrons haven been put for the initial excited transition, the state translates to HOMO + from_state
-    - `beta`: corresponds to $(k_B T)^{-1}$
-    """
-
-    hamiltonian: jax.Array
-    coulomb: jax.Array
-    rho_0: jax.Array
-    rho_stat: jax.Array
-    energies: jax.Array
-    eigenvectors: jax.Array
-    positions: jax.Array
-    unique_ids: list[str]
-    ids: jax.Array
-    sublattice_ids: jax.Array
-    eps: float
-    homo: int
-    electrons: int
-    from_state: int
-    to_state: int
-    excited_electrons: int
-    beta: float
-    spin_degeneracy: int
-    transitions: TransitionsDict
-
-
-class LatticeType(Enum):
-    CHAIN = auto()
-    SSH = auto()
-    HONEYCOMB = auto()
-
-
-class LatticeEdge(Enum):
-    NONE = auto()
-    ARMCHAIR = auto()
-    ZIGZAG = auto()
-
-
-## GEOMETRY
-# TODO: this is somewhat broken
-@struct.dataclass
-class Chain:
-    """A chain.
-
-    - `length`: length of the chain
-    """
-
-    length: float
-
-    def _layout(self, lattice_constant: float) -> jax.Array:
-        n = int(self.length / lattice_constant)
-        return jnp.expand_dims(jnp.arange(2 * n), 1) % 2
-
-
-@struct.dataclass
-class Triangle:
-    """An equilateral triangle.
-
-    - `base`: Base length of the triangle
-    """
-
-    base: float
-
-    def _layout(self, lattice_constant: float) -> jax.Array:
-        n = int(self.base / lattice_constant)
-
-        size = 2 * n + 1
-        shape = np.zeros((size, n), dtype=bool)
-        for i in range(n):
-            shape[n + i, i::2] = True
-            shape[n - i, i::2] = True
-        return jnp.array(shape)
-
-
-@struct.dataclass
-class Hexagon:
-    """A hexagon.
-
-    - `side_length`: Side length of the hexagon
-    """
-
-    side_length: float
-
-    def _layout(self, lattice_constant: float) -> jax.Array:
-        n = int(self.side_length / lattice_constant)
-
-        width = 2 * n + 1
-        height = 4 * n - 2
-        shape = np.zeros((height, width))
-
-        for i in range(n):
-            shape[i : (height - i) : 2, n + i] = True
-            shape[i : (height - i) : 2, n - i] = True
-
-        return jnp.array(shape)
-
-
-@struct.dataclass
-class Rhomboid:
-    """A rhomboid.
-
-    - `x`: side length in x-diretion
-    - `y`: side length in y-diretion
-    """
-
-    x: float
-    y: float
-
-    def _layout(self, lattice_constant: float) -> jax.Array:
-        x, y = int(self.x / lattice_constant), int(self.y / lattice_constant)
-        shape = np.zeros((2 * (x + y), y), dtype=bool)
-        for i in range(y):
-            upper = 2 * x + i
-            shape[i:upper:2, i] = True
-        return jnp.array(shape)
-
-
-@struct.dataclass
-class Rectangle:
-    """A rectangle.
-
-    - `x`: side length in x-diretion
-    - `y`: side length in y-diretion
-    """
-
-    x: float
-    y: float
-
-    def _layout(self, lattice_constant: float) -> jax.Array:
-        x, y = int(self.x / lattice_constant), int(self.y / lattice_constant)
-
-        height = y * 2
-        width = 2 * x - 1
-        shape = np.zeros((height, width), dtype=bool)
-        for i in range(height):
-            shape[i, (i % 2) : width : 2] = True
-        return jnp.array(shape)
-
-
-class Lattice:
-    """Monoatomic lattice cut.
-
-    A triangular graphene lattice cut would be described as follows:
-
-    ``` py
-        graphene = granad.Lattice(
-                   shape=triangle,
-                   lattice_edge=granad.LatticeEdge.ARMCHAIR,
-                   lattice_type=granad.LatticeType.HONEYCOMB,
-                   lattice_constant=2.46,
-                  )
-    ```
-    - `shape`: a boolean array indicating the shape of the structure
-    - `lattice_type`: lattice type
-    - `lattice_edge`: edge type
-    - `lattice_constant`: lattice constant of the infinite lattice
-    - `shift`: spacial shift to apply to all coordinates
-    """
-
-    ## entries in this dict follow the scheme:
-    ## (lattice_type, edge_type) : (confguration, scale, lattice basis, unit cell)
-    ## the first entry is a redundant unit cell, which is copied to produce the correct edge type
-    ## the the second entry refers to the scaling applied to a layout array
-    ## the last two entries in the tuple refer to the infinite lattice
-
-    _structures = {
-        (LatticeType.HONEYCOMB, LatticeEdge.ARMCHAIR): (
-            jnp.array(
-                [
-                    [0, 0, 0],
-                    [1, 0, 0],
-                    [3 / 2, -jnp.sqrt(3) / 2, 0],
-                    [1, -2 * jnp.sqrt(3) / 2, 0],
-                    [0, -2 * jnp.sqrt(3) / 2, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                ]
-            ),
-            jnp.array([3 / 2, 3 / 2 * jnp.sqrt(3)]),
-            jnp.array(
-                [
-                    [0, 0, 0],
-                    [1, 0, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                ]
-            ),
-            jnp.array(
-                [
-                    [0, 0, 0],
-                    [1, 0, 0],
-                    [3 / 2, -jnp.sqrt(3) / 2, 0],
-                    [1, -2 * jnp.sqrt(3) / 2, 0],
-                    [0, -2 * jnp.sqrt(3) / 2, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                ]
-            ),
-        ),
-        (LatticeType.HONEYCOMB, LatticeEdge.ZIGZAG): (
-            jnp.array(
-                [
-                    [-jnp.sqrt(3) / 2, 1 / 2, 0],
-                    [0, 1, 0],
-                    [jnp.sqrt(3) / 2, 1 / 2, 0],
-                    [-jnp.sqrt(3) / 2, -1 / 2, 0],
-                    [0, -1, 0],
-                    [jnp.sqrt(3) / 2, -1 / 2, 0],
-                ]
-            ),
-            jnp.array([1 / 2 * jnp.sqrt(3), 3 / 2]),
-            jnp.array(
-                [
-                    [0, 0, 0],
-                    [1, 0, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                ]
-            ),
-            jnp.array(
-                [
-                    [0, 0, 0],
-                    [1, 0, 0],
-                    [3 / 2, -jnp.sqrt(3) / 2, 0],
-                    [1, -2 * jnp.sqrt(3) / 2, 0],
-                    [0, -2 * jnp.sqrt(3) / 2, 0],
-                    [-1 / 2, -jnp.sqrt(3) / 2, 0],
-                ]
-            ),
-        ),
-        (LatticeType.CHAIN, LatticeEdge.NONE): (
-            jnp.array([[0.0, 0, 0]]),
-            jnp.array([0.5, 1.0]),
-            jnp.array(
-                [
-                    [1.0, 0.0, 0.0],
-                ]
-            ),
-            jnp.array(
-                [
-                    [0, 0, 0],
-                ]
-            ),
-        ),
-        (LatticeType.SSH, LatticeEdge.NONE): (
-            jnp.array([[0.0, 0, 0], [0.3, 0, 0]]),
-            jnp.array([0.5, 1.0]),
-            jnp.array(
-                [
-                    [1.0, 0.0, 0.0],
-                ]
-            ),
-            jnp.array([[0, 0, 0], [0.3, 0.0, 0.0]]),
-        ),
-    }
-
-    def __init__(
-        self,
-        shape: Union[Chain, Triangle, Hexagon, Rhomboid, Rectangle],
-        lattice_type: LatticeType,
-        lattice_edge: LatticeEdge,
-        lattice_constant: float,
-        shift=jnp.zeros(3),
-    ):
-        """Initialization"""
-
-        try:
-            site_distance = self._site_distance(lattice_constant, lattice_type)
-            self.configuration, self.scale, self.lattice_basis, self.unit_cell = [
-                x * site_distance
-                for x in self._structures[(lattice_type, lattice_edge)]
-            ]
-        except KeyError:
-            raise Exception(
-                f"Lattice {lattice_type} with edge {lattice_edge} is not supported!"
-            )
-
-        self.shift = shift
-        self.shape = shape._layout(lattice_constant)
-
-        if not (lattice_type, lattice_edge) in self._structures.keys():
-            raise Exception(f"Lattice {lattice_type} not implemented!")
-
-    @staticmethod
-    def _site_distance(lattice_constant: float, lattice_type: LatticeType) -> float:
-        if lattice_type == LatticeType.HONEYCOMB:
-            return lattice_constant / jnp.sqrt(3)
-        else:
-            return lattice_constant
-
-    def _orbitals(self, orbital_id: str, occupation: int = 1) -> list[Orbital]:
-        """Returns a list of orbitals in a geometric arrangement."""
-
-        shape, shift, configuration, scale = (
-            self.shape,
-            self.shift,
-            self.configuration,
-            self.scale,
-        )
-
-        # TODO: fix sublattice_id
-        def add_orbitals(position: jax.Array):
-            return [
-                Orbital(
-                    position=jnp.array(
-                        list(scale * position + displacement[:2] + shift[:2])
-                        + [shift[2] + displacement[2]]
-                    ),
-                    orbital_id=orbital_id,
-                    occupation=occupation,
-                    sublattice_id=i % 2,
-                )
-                for i, displacement in enumerate(configuration)
-            ]
-
-        orbitals = reduce(
-            lambda x, y: x + y, map(lambda pos: add_orbitals(pos), jnp.argwhere(shape))
-        )
-        return [
-            orbitals[i]
-            for i in jnp.unique(
-                jnp.round(jnp.array([orb.position for orb in orbitals]), 8),
-                axis=0,
-                return_index=True,
-            )[1]
-        ]
-
-
-@struct.dataclass
-class Spot:
-    """Single point in 3D space. Can represent isolated adatoms.
-
-    - `position`: point, e.g. `[0.0, 1.0, 2.0]`
-    """
-
-    position: list[float]
-
-
-@struct.dataclass
-class DistanceCoupling:
-    """Distance-dependent coupling.
-
-    For example, a model direct channel repulsion of $\\frac{1}{r}$ between
-    two orbitals identified by the strings `"X"` and `"Y"` can be modelled
-    as follows:
-
-    ``` py
-    coulomb_repulsion = SpotCoupling(
-                        orbital_id1 = "X", # this id must match the id given to stack.add
-                        orbital_id2 = "Y", # this id must match the id given to stack.add
-                        coupling = lambda r : 1/r + 0j
-         )
-    ```
-
-    - `orbital_id1`: orbital id
-    - `orbital_id2`: orbital id
-    - `coupling_function`: coupling function, must return a **complex number**
-    """
-
-    orbital_id1: str
-    orbital_id2: str
-    coupling_function: Callable[[float], complex]
-
-
-@struct.dataclass
-class SpotCoupling:
-    """Onsite coupling. Can hold the coupling between two orbitals at an adatom.
-
-    For example, the onsite energy of a single adatom orbital identified by the string `"A"` can be modelled
-    as follows:
-
-    ``` py
-       onsite_adatom = SpotCoupling(
-       orbital_id1 = "A", # this id must match the id given to stack.add
-       orbital_id2 = "A",
-       coupling = 0.0
-       )
-    ```
-
-    - `orbital_id1`: orbital id
-    - `orbital_id2`: orbital id
-    - `coupling`: coupling strength
-    """
-
-    orbital_id1: str
-    orbital_id2: str
-    coupling: float
-
-
-@struct.dataclass
-class LatticeCoupling:
-    """Coupling in a lattice material.
-
-    Constructor accepts a lattice object and a list of couplings.
-    The i-th entry in the couplings-list determines the coupling to the i-th neighbor,
-    i.e. `couplings[0]` is the onsite coupling, `couplings[2]` is the NNN-coupling.
-
-    For example, the Coulomb coupling in spin-polarized graphene with an onsite strength
-    of `16.522`, nearest-neighbor interaction of `8.64`, next-to-nearest neighbor
-    interaction of `5.333` and a classical coupling for orbitals further away can be modelled
-    as follows:
-
-    ``` py
-       coulomb_graphene = LatticeCoupling(
-       orbital_id1 = "pz", # this id must match the id given to stack.add
-       orbital_id2 = "pz",
-       lattice = graphene,
-       couplings = [16.522, 8.64, 5.333],
-       coupling_function = lambda d: 14.399 / d + 0j,
-       )
-    ```
-
-    - `orbital_id1`: orbital id
-    - `orbital_id2`: orbital id
-    - `lattice`: lattice material
-    - `couplings`: couplings
-    - `coupling_function`: function to apply to all neighbours not covered by couplings
-    """
-
-    orbital_id1: str
-    orbital_id2: str
-    lattice: Lattice
-    couplings: list[float]
-    coupling_function: Callable[[float], complex] = lambda x: 0j
-
-
-@struct.dataclass
-class LatticeSpotCoupling:
-    """Determines coupling between a lattice material and an isolated spot.
-
-    Constructor accepts a list of couplings.
-    The i-th entry in the couplings-list determines the coupling between the isolated spot
-    and its i-th neighbor in the lattice cut,
-    i.e. `couplings[0]` is the onsite coupling, `couplings[2]` is the NNN-coupling.
-
-    For example, the Coulomb coupling of a single adatom orbital by the string `"A"` to
-    a flake orbital identified by the string `"pz"` with an onsite strength
-    of `16.522`, nearest-neighbor interaction of `8.64`, next-to-nearest neighbor
-    interaction of `5.333` and a classical coupling for orbitals further away can be modelled
-    as follows:
-
-    ``` py
-       coulomb_graphene = LatticeSpotCoupling(
-       lattice_id = "pz", # this id must match the id given to stack.add
-       spot_id = "A", # this id must match the id given to stack.add
-       couplings = [16.522, 8.64, 5.333],
-       coupling_function = lambda d: 14.399 / d + 0j,
-       )
-    ```
-
-    - `lattice_id`: bulk orbital id
-    - `spot_id`: isolated orbital id
-    - `couplings`: couplings list
-    - `coupling_function`: function to apply to all neighbours not covered by couplings
-    """
-
-    lattice_id: str
-    spot_id: str
-    couplings: list[float]
-    coupling_function: Callable[[float], complex] = lambda x: 0j
-
-
-def gaussian_coupling(sigma, mu, couplings):
-    """Helper function for "almost-neighbor" coupling"""
-    return lambda x: sum(
-        couplings[i] * jnp.exp(-((x - mu[i]) ** 2) / sigma**2)
-        for i in range(len(couplings))
-    )
-
-
-class StackBuilder:
-    """Incremental specification of a composite nanomaterial.
-
-    Orbitals are added one by one and their TB-hopping and coulomb interactions specified.
-    """
-
-    def __init__(self):
-        """Initialization."""
-        self.orbitals: list[Orbital] = []
-        self.hopping: dict[tuple[str, str], Callable[[float], complex]] = {}
-        self.coulomb: dict[tuple[str, str], Callable[[float], complex]] = {}
-
-        self.spots: int = 0
-        self.lattices: int = 0
-
-    @property
-    def orbital_ids(self):
-        return list(set([orb.orbital_id for orb in self.orbitals]))
-
-    def add(self, orbital_id: str, material: Union[Spot, Lattice], occupation: int = 1):
-        """Adds new orbitals to the stack.
-
-        - `orbital_id`: new orbital to be added
-        - `material`: material, an isolated system or a lattice cut
-        - `occupation`: number of electrons carried by the orbital
-        """
-
-        if isinstance(material, Spot):
-            # for a single site, add the orbital at that position
-            self.orbitals.append(Orbital(orbital_id, material.position, occupation))
-
-        elif isinstance(material, Lattice):
-            # for a lattice, append a list of orbitals in the specified configuartion
-            self.orbitals += material._orbitals(orbital_id, occupation)
-        else:
-            raise TypeError(f"Incorrect material type {material}!")
-
-    def _set_coupling(
-        self,
-        coupling_dict: dict,
-        coupling: Union[
-            DistanceCoupling, SpotCoupling, LatticeCoupling, LatticeSpotCoupling
-        ],
-    ):
-        if isinstance(coupling, DistanceCoupling):
-            coupling_dict[(coupling.orbital_id1, coupling.orbital_id2)] = jax.jit(
-                coupling.coupling_function
-            )
-        elif isinstance(coupling, SpotCoupling):
-            coupling_dict[(coupling.orbital_id1, coupling.orbital_id2)] = jax.jit(
-                lambda d: coupling.coupling + 0j
-            )
-        elif isinstance(coupling, LatticeCoupling):
-            coupling_dict[(coupling.orbital_id1, coupling.orbital_id2)] = (
-                self._lattice_coupling(coupling)
-            )
-        elif isinstance(coupling, LatticeSpotCoupling):
-            lattice_pos = self.get_positions(coupling.lattice_id)
-            spot_pos = self.get_positions(coupling.spot_id)
-            coupling_dict[(coupling.lattice_id, coupling.spot_id)] = (
-                self._lattice_spot_coupling(
-                    coupling, lattice_pos=lattice_pos, spot_pos=spot_pos
-                )
-            )
-        else:
-            raise TypeError(f"Incorrect coupling type {coupling}!")
-
-    def set_hopping(
-        self,
-        hopping: Union[
-            DistanceCoupling, SpotCoupling, LatticeCoupling, LatticeSpotCoupling
-        ],
-    ):
-        """Specifies the TB-hopping between two types of orbitals.
-
-        - `hopping`: hopping strength
-        """
-
-        self._set_coupling(self.hopping, hopping)
-
-    def set_coulomb(
-        self,
-        coulomb: Union[
-            DistanceCoupling, SpotCoupling, LatticeCoupling, LatticeSpotCoupling
-        ],
-    ):
-        """Specifies the Coulomb interaction between two types of orbitals.
-
-        - `coulomb`: coulomb coupling
-        """
-
-        self._set_coupling(self.coulomb, coulomb)
-
-    def get_positions(self, orbital_id: str = None) -> jax.Array:
-        """Gets the positions in the stack.
-
-                - `orbital_id`: If supplied, only get the positions of this orbital
-
-        **Returns:**
-
-         $N \\times 3$ - array, where $N$ is the number of orbitals
-        """
-
-        if orbital_id is None:
-            return jnp.array([x.position for x in self.orbitals])
-        else:
-            return jnp.array(
-                [x.position for x in self.orbitals if x.orbital_id == orbital_id]
-            )
-
-    def show2D(self):
-        """Plots all orbitals in 2D"""
-
-        norm = Normalize(0, len(self.orbital_ids))
-
-        fig, ax = plt.subplots(1, 1)
-        for orb in self.orbital_ids:
-            positions = self.get_positions(orb)
-            color = norm(self.orbital_ids.index(orb)) * jnp.ones_like(positions[:, 0])
-            ax.scatter(
-                *zip(*positions[:, :2]),
-                s=40,
-                c=color,
-                label=orb,
-            )
-
-        pos = self.get_positions()
-        for i in range(len(self.orbitals)):
-            ax.annotate(
-                str(i),
-                (
-                    pos[i, 0],
-                    pos[i, 1],
-                ),
-            )
-        ax.axis("equal")
-        plt.legend()
-        plt.show()
-
-    def show3D(self):
-        """Plots all orbitals in 3D"""
-
-        norm = Normalize(0, len(self.orbital_ids))
-
-        fig = plt.figure()
-        ax = fig.add_subplot(111, projection="3d")
-        for orb in self.orbital_ids:
-            positions = self.get_positions(orb)
-            color = norm(self.orbital_ids.index(orb)) * jnp.ones_like(positions[:, 0])
-            ax.scatter(
-                *zip(*positions[:, :2]),
-                zs=positions[:, 2],
-                s=40,
-                c=color,
-                label=orb,
-            )
-
-        pos = self.get_positions()
-        for i in range(len(self.orbitals)):
-            ax.text(*pos[i, :], str(i), "x")
-
-        plt.legend()
-        plt.show()
-
-    def _ensure_combinations(self, coupling_dict, name):
-        for id1 in self.orbital_ids:
-            for id2 in self.orbital_ids:
-                if not (
-                    (id1, id2) in coupling_dict.keys()
-                    or (id2, id1) in coupling_dict.keys()
-                ):
-                    raise Exception(
-                        f"Not all {name} combinations specified, missing {(id1,id2)}!"
-                    )
-
-    @property
-    def sublattice_ids(self):
-        return [orb.sublattice_id for orb in self.orbitals]
-
-    # TODO: does the field_func make sense?
-    def get_stack(
-        self,
-        field_func: Callable[[float], jax.Array] = lambda x: 0j,
-        from_state: int = 0,
-        to_state: int = 0,
-        excited_electrons: int = 1,
-        doping: int = 0,
-        beta: float = jnp.inf,
-        eps: float = 1e-5,
-        spin_degenerate: bool = True,
-        transitions: TransitionsDict = None,
-        pingback: bool = True,
-        energies_only: bool = False,
-    ) -> Stack:
-        """Get stack object for numerical simulations.
-
-                - `field_func`: position-dependent additive term for hamiltonian (e.g. electric field)
-                - `from_state`: optional, the state from which electrons to take for the initial excited transition, the state translates to HOMO - from_state
-                - `to_state`: optional, the state into which electrons to put for the initial excited transition, the state translates to HOMO + from_state
-                - `doping`: optional, how many doping electrons to throw into the stack
-                - `beta`: thermodynamic beta $\\frac{1}{k_B T}$
-                - `eps`: optional, default `1e-5`: numerical threshold. if two energies E1,E2 fullfill abs(E1-E2) < eps, they are considered degenerate
-
-        **Returns:**
-
-         object
-        """
-
-        self._ensure_combinations(self.hopping, "hopping")
-        if energies_only:
-            self.coulomb = {
-                combination: lambda x: 0.0 for combination in self.hopping.keys()
-            }
-            self._ensure_combinations(self.coulomb, "coulomb")
-
-        stack = _stack(
-            self.orbitals,
-            self.hopping,
-            self.coulomb,
-            field_func,
-            from_state,
-            to_state,
-            excited_electrons,
-            doping,
-            beta,
-            jnp.array(self.sublattice_ids),
-            eps,
-            2.0 if spin_degenerate else 1.0,
-            transitions,
-        )
-        if pingback:
-            print(f"Built stack with {stack.electrons} electrons")
-
-        return stack
-
-    @staticmethod
-    def _lattice_spot_coupling(
-        lattice_spot_coupling: LatticeSpotCoupling,
-        *,
-        lattice_pos: jax.Array,
-        spot_pos: jax.Array,
-        tolerance: float = 1e-5,
-    ):
-        couplings = jnp.array([x + 0j for x in lattice_spot_coupling.couplings])
-
-        distances = jnp.sort(
-            jnp.unique(jnp.round(jnp.linalg.norm(lattice_pos - spot_pos, axis=1), 8))
-        )[: len(couplings)]
-
-        def inner(d):
-            return jax.lax.cond(
-                jnp.min(jnp.abs(d - distances)) < tolerance,
-                lambda x: couplings[jnp.argmin(jnp.abs(x - distances))],
-                lattice_spot_coupling.coupling_function,
-                d,
-            )
-
-        return jax.jit(inner)
-
-    @staticmethod
-    def _lattice_coupling(
-        lattice_coupling: LatticeCoupling,
-        tolerance: float = 1e-5,
-    ):
-        """Takes in some orbitals and geometry. Returns a function that is suitable as a value in an InputDict. The returned function takes a distance and returns the appropriate nearest neighbour coupling.
-
-        - `lattice_coupling`:
-        - `tolerance`:
-
-        **Returns:**
-        A function that is suitable as a value in an InputDict
-        """
-
-        lattice_basis = lattice_coupling.lattice.lattice_basis
-        unit_cell = lattice_coupling.lattice.unit_cell
-        couplings = jnp.array([x + 0j for x in lattice_coupling.couplings])
-
-        positions = []
-        for i in range(len(couplings)):
-            for pos in unit_cell:
-                for v in lattice_basis:
-                    positions.append(pos + i * v)
-        positions = jnp.array(positions)
-        distances = jnp.unique(
-            jnp.round(jnp.linalg.norm(positions - positions[:, None, :], axis=2), 8)
-        )[: len(couplings)]
-
-        def inner(d):
-            return jax.lax.cond(
-                jnp.min(jnp.abs(d - distances)) < tolerance,
-                lambda x: couplings[jnp.argmin(jnp.abs(x - distances))],
-                lattice_coupling.coupling_function,
-                d,
-            )
-
-        return jax.jit(inner)
-
-
-# density matrices
 def fermi(e, beta, mu):
     return 1 / (jnp.exp(beta * (e - mu)) + 1)
-
 
 def density_matrix(
     energies,
@@ -973,91 +148,6 @@ def _density_aufbau(
     return jnp.diag(occupation) / electrons, homo
 
 
-def _stack(
-    orbs: list[Orbital],
-    hopping: InputDict,
-    coulomb: InputDict,
-    field_func: Callable[[float], jax.Array],
-    from_state: int,
-    to_state: int,
-    excited_electrons: int,
-    doping: int,
-    beta: float,
-    sublattice_ids: jax.Array,
-    eps: float,
-    spin_degeneracy: int,
-    transitions: TransitionsDict,
-) -> Stack:
-    """Takes a list of orbitals and two dictionaries specifying the hopping rates and coulomb interactions and produces a stack that holds the state of the system before it is propagated in time (e.g. the eigenenergies and density matrix).
-
-        - `orbs`: list of orbitals
-        - `hopping`: input dictionary for the hopping rates
-        - `coulomb`: input dictionary for the coulomb couplings
-        - `field_func`: position-dependent additive term for hamiltonian (e.g. electric field)
-        - `from_state`: optional, the state from which electrons to take for the initial excited transition, the state translates to HOMO - from_state
-        - `to_state`: optional, the state into which electrons to put for the initial excited transition, the state translates to HOMO + from_state
-        - `doping`: optional, how many doping electrons to throw into the stack
-        - `eps`: optional, default 1e-5: numerical threshold. if two energies E1,E2 fullfill abs(E1-E2) < eps, they are considered degenerate
-
-    **Returns:**
-
-     object
-    """
-
-    pos, ids, func, unique_ids = _unpack(orbs, hopping, coulomb)
-    idxs = jnp.arange(pos.shape[0])
-
-    hamiltonian, coulomb = _hamiltonian_coulomb(pos, ids, func)(idxs, idxs)
-
-    electrons = sum(x.occupation for x in orbs) + doping
-
-    eigenvectors, energies = jax.lax.linalg.eigh(hamiltonian + field_func(pos))
-
-    # TODO check same length
-    from_state = jnp.array([from_state] if isinstance(from_state, int) else from_state)
-    to_state = jnp.array([to_state] if isinstance(to_state, int) else to_state)
-    excited_electrons = jnp.array(
-        [excited_electrons] if isinstance(excited_electrons, int) else excited_electrons
-    )
-
-    eq_arr = jnp.array([0])
-    rho_0, homo = _density_matrix(
-        energies,
-        electrons,
-        spin_degeneracy,
-        eps,
-        from_state,
-        to_state,
-        excited_electrons,
-        beta,
-    )
-    rho_stat, _ = _density_matrix(
-        energies, electrons, spin_degeneracy, eps, eq_arr, eq_arr, eq_arr, beta
-    )
-
-    return Stack(
-        hamiltonian + field_func(pos),
-        coulomb,
-        rho_0,
-        rho_stat,
-        energies,
-        eigenvectors,
-        pos,
-        unique_ids,
-        ids,
-        sublattice_ids,
-        eps,
-        homo,
-        electrons,
-        from_state,
-        to_state,
-        excited_electrons,
-        beta,
-        spin_degeneracy,
-        transitions,
-    )
-
-
 ## DIPOLE TRANSITIONS
 def dipole_transitions(
     stack: Stack,
@@ -1139,7 +229,7 @@ def dipole_transitions(
         )
         - stack.positions[indices[:, 0, :].flatten(), :]
     )
-
+p
     # array of shape positions x orbitals x 3, with entries r_point_charge[i, o, :] = (r_o - r_i)/|r_o - r_i|^3
     r_point_charge = jnp.nan_to_num(
         vec_r / jnp.expand_dims(jnp.linalg.norm(vec_r, axis=2) ** 3, 2),
@@ -1149,8 +239,6 @@ def dipole_transitions(
 
     return jax.jit(inner)
 
-
-## DISSIPATION
 def relaxation(tau: float) -> Callable:
     """Function for modelling dissipation according to the relaxation approximation.
 
@@ -1198,8 +286,6 @@ def lindblad(stack, gamma, saturation):
 
     return inner
 
-
-## self-consistency
 def get_self_consistent(
     stack: Stack, iterations: int = 50, mix: float = 0.05, accuracy: float = 1e-6
 ) -> Stack:
@@ -1290,176 +376,6 @@ def get_self_consistent(
         homo=homo,
     )
 
-
-## ELECTRIC FIELDS
-def electric_field(
-    amplitudes: list[float],
-    frequency: float,
-    positions: jax.Array,
-    k_vector: list[float] = [0.0, 0.0, 1.0],
-):
-    """Function for computing time-harmonic electric fields.
-
-        - `amplitudes`: electric field amplitudes in xyz-components
-        - `frequency`: frequency
-        - `positions`: position for evaluation
-
-    **Returns:**
-
-    -compiled closure that computes the electric field as a functon of time
-    """
-    static_part = jnp.expand_dims(jnp.array(amplitudes), 1) * jnp.exp(
-        -1j * jnp.pi / 2 + 1j * positions @ jnp.array(k_vector)
-    )
-    return jax.jit(lambda t: jnp.exp(1j * frequency * t) * static_part)
-
-
-def electric_field_with_ramp_up(
-    amplitudes: list[float],
-    frequency: float,
-    positions: jax.Array,
-    ramp_duration: float,
-    time_ramp: float,
-    k_vector: list[float] = [0.0, 0.0, 1.0],
-):
-    """Function for computing ramping up time-harmonic electric fields.
-
-        - `amplitudes`: electric field amplitudes in xyz-components
-        - `frequency`: frequency
-        - `positions`: positions for evaluation
-        - `ramp_duration`: specifies how long does the electric field ramps up
-        - `time_ramp`: specifies time at which the field starts to ramp up
-
-    **Returns:**
-
-    -compiled closure that computes the electric field as a functon of time
-    """
-    static_part = jnp.expand_dims(jnp.array(amplitudes), 1) * jnp.exp(
-        -1j * positions @ jnp.array(k_vector)
-    )
-    p = 0.99
-    ramp_constant = 2 * jnp.log(p / (1 - p)) / ramp_duration
-    return jax.jit(
-        lambda t: static_part
-        * jnp.exp(1j * frequency * t)
-        / (1 + 1.0 * jnp.exp(-ramp_constant * (t - time_ramp)))
-    )
-
-
-def electric_field_pulse(
-    amplitudes: list[float],
-    frequency: float,
-    positions: jax.Array,
-    peak: float,
-    fwhm: float,
-    k_vector: list[float] = [0.0, 0.0, 1.0],
-):
-    """Function for computing temporally located time-harmonics electric fields. The pulse is implemented as a temporal Gaussian.
-
-        - `amplitudes`: electric field amplitudes in xyz-components
-        - `frequency`: frequency of the electric field
-        - `positions`: positions where the electric field is evaluated
-        - `peak`: time where the pulse reaches its peak
-        - `fwhm`: full width at half maximum
-
-    **Returns:**
-
-    -compiled closure that computes the electric field as a functon of time
-    """
-
-    static_part = jnp.expand_dims(jnp.array(amplitudes), 1) * jnp.exp(
-        -1j * positions @ jnp.array(k_vector)
-    )
-    sigma = fwhm / (2.0 * jnp.sqrt(jnp.log(2)))
-    return jax.jit(
-        lambda t: static_part
-        * jnp.exp(-1j * jnp.pi / 2 + 1j * frequency * (t - peak))
-        * jnp.exp(-((t - peak) ** 2) / sigma**2)
-    )
-
-
-def dos(stack: Stack, omega: float, broadening: float = 0.1) -> jax.Array:
-    """IP-DOS of a nanomaterial stack.
-
-    - `stack`: a stack object
-    - `omega`: frequency
-    - `broadening`: numerical brodening parameter to replace Dirac Deltas with
-    """
-
-    broadening = 1 / broadening
-    prefactor = 1 / (jnp.sqrt(2 * jnp.pi) * broadening)
-    gaussians = jnp.exp(-((stack.energies - omega) ** 2) / 2 * broadening**2)
-    return prefactor * jnp.sum(gaussians)
-
-
-def ldos(
-    stack: Stack, omega: float, site_index: int, broadening: float = 0.1
-) -> jax.Array:
-    """IP-LDOS of a nanomaterial stack.
-
-    - `stack`: a stack object
-    - `omega`: frequency
-    - `site_index`: the site index to evaluate the LDOS at
-    - `broadening`: numerical brodening parameter to replace Dirac Deltas with
-    """
-
-    broadening = 1 / broadening
-    weight = jnp.abs(stack.eigenvectors[site_index, :]) ** 2
-    prefactor = 1 / (jnp.sqrt(2 * jnp.pi) * broadening)
-    gaussians = jnp.exp(-((stack.energies - omega) ** 2) / 2 * broadening**2)
-    return prefactor * jnp.sum(weight * gaussians)
-
-
-## TIME PROPAGATION
-def quadrupole_operator(stack):
-    dip = position_operator(stack)
-    term = jnp.einsum("ijk,jlm->ilkm", dip, dip)
-    diag = jnp.einsum("ijk,jlk->il", dip, dip)
-    diag = jnp.einsum("ij,kl->ijkl", diag, jnp.eye(term.shape[-1]))
-    return 3 * term - diag
-
-
-def position_operator(stack):
-    N = stack.positions.shape[0]
-    pos = jnp.zeros((N, N, 3))
-    for i in range(3):
-        pos = pos.at[:, :, i].set(jnp.diag(stack.positions[:, i] / 2))
-    if not stack.transitions is None:
-        for key, value in stack.transitions.items():
-            value = jnp.array(value)
-            i, j = indices(stack, key[0]), indices(stack, key[1])
-            k = value.nonzero()[0]
-            pos = pos.at[i, j, k].set(value[k])
-    return pos + jnp.transpose(pos, (1, 0, 2))
-
-
-def velocity_operator(stack):
-    if stack.transitions is None:
-        x_times_h = jnp.einsum("ij,ir->ijr", stack.hamiltonian, stack.positions)
-        h_times_x = jnp.einsum("ij,jr->ijr", stack.hamiltonian, stack.positions)
-    else:
-        positions = position_operator(stack)
-        x_times_h = jnp.einsum("kj,ikr->ijr", stack.hamiltonian, positions)
-        h_times_x = jnp.einsum("ik,kjr->ijr", stack.hamiltonian, positions)
-    return -1j * (x_times_h - h_times_x)
-
-
-def fraction_periodic(signal, threshold=1e-2):
-
-    # signal is periodic => abs(signal) is periodic
-    cum_sum = jnp.abs(signal).cumsum()
-
-    # cumulative mean of periodic signal is constant
-    cum_mean = cum_sum / jnp.arange(1, len(signal) + 1)
-
-    # if cumulative mean doesn't move anymore, we have a lot of periodic signal
-    med = jnp.median(cum_mean)
-    deviation = jnp.abs(med - cum_mean) / med
-
-    # approximate admixture of periodic signal
-    return (deviation < threshold).sum() / len(signal)
-
-
 # TODO: units
 def evolution(
     stack: Stack,
@@ -1472,7 +388,29 @@ def evolution(
     stepsize_controller=diffrax.PIDController(rtol=1e-8, atol=1e-8),
     spatial=False,
     add_induced=False,
+    custom = False
 ):
+
+    def integrate_custom(rho, time):
+        e_field, delta_rho = field(time), rho - rho_stat
+        charge = -jnp.diag(delta_rho) * stack.electrons
+        p_ext = jnp.sum(stack.positions * e_field.real.T, axis=1)
+        p_ind = coulomb @ charge
+        h_total = transition(charge, stack.hamiltonian, e_field) + jnp.diag(
+            p_ext - p_ind
+        )
+        if dissipation:
+            return (
+                rho
+                - 1j * dt * (h_total @ rho - rho @ h_total)
+                + dt * dissipation(rho, rho_stat),
+                postprocess(rho) if postprocess else rho,
+            )
+        else:
+            return (
+                rho - 1j * dt * (h_total @ rho - rho @ h_total),
+                postprocess(rho) if postprocess else rho,
+            )
 
     def rhs_uniform(time, rho, args):
         e_field, delta_rho = field(time), rho - rho_stat
@@ -1509,10 +447,22 @@ def evolution(
         rhs = rhs_uniform
         transition = dipole_transitions(stack, add_induced)
 
+    if custom:
+
+        dt = time[1] - time[0]
+        coulomb = stack.coulomb * coulomb_strength
+        rho_stat = stack.eigenvectors @ stack.rho_stat @ stack.eigenvectors.conj().T
+        transition = dipole_transitions(stack, add_induced=add_induced)
+        rho, rhos = jax.lax.scan(
+            integrate_custom, stack.eigenvectors @ stack.rho_0 @ stack.eigenvectors.conj().T, time
+        )
+
+        return rhos
+
     term = diffrax.ODETerm(rhs)
     rho_init = stack.eigenvectors @ stack.rho_0 @ stack.eigenvectors.conj().T
     saveat = diffrax.SaveAt(ts=time if saveat is None else saveat)
-    sol = diffrax.diffeqsolve(
+    return diffrax.diffeqsolve(
         term,
         solver,
         t0=time[0],
@@ -1522,153 +472,7 @@ def evolution(
         saveat=saveat,
         stepsize_controller=stepsize_controller,
     )
-    return (
-        stack.replace(
-            rho_0=stack.eigenvectors.conj().T @ sol.ys[-1] @ stack.eigenvectors
-        ),
-        sol,
-    )
 
-
-def evolution_old(
-    stack: Stack,
-    time: jax.Array,
-    field: FieldFunc,
-    dissipation: DissipationFunc = None,
-    coulomb_strength: float = 1.0,
-    postprocess: Callable[[jax.Array], jax.Array] = None,
-    add_induced=False,
-) -> tuple[Stack, jax.Array]:
-    """Propagate a stack forward in time.
-
-        - `stack`: stack object
-        - `time`: time axis
-        - `field`: electric field function
-        - `dissipation`: dissipation function
-        - `coulomb_strength`: scaling factor applied to coulomb interaction strength
-        - `dipole_transition`: a function describing dipolar transition radiation in the stack
-        - `postprocess`: a function applied to the density matrix after each time step
-
-    **Returns:**
-
-    (stack with rho_0 set to the current state, array containing all results from postprocess at all timesteps)
-    """
-
-    def integrate(rho, time):
-        e_field, delta_rho = field(time), rho - rho_stat
-        charge = -jnp.diag(delta_rho) * stack.electrons
-        p_ext = jnp.sum(stack.positions * e_field.real.T, axis=1)
-        p_ind = coulomb @ charge
-        h_total = transition(charge, stack.hamiltonian, e_field) + jnp.diag(
-            p_ext - p_ind
-        )
-        if dissipation:
-            return (
-                rho
-                - 1j * dt * (h_total @ rho - rho @ h_total)
-                + dt * dissipation(rho, rho_stat),
-                postprocess(rho) if postprocess else rho,
-            )
-        else:
-            return (
-                rho - 1j * dt * (h_total @ rho - rho @ h_total),
-                postprocess(rho) if postprocess else rho,
-            )
-
-    dt = time[1] - time[0]
-    coulomb = stack.coulomb * coulomb_strength
-    rho_stat = stack.eigenvectors @ stack.rho_stat @ stack.eigenvectors.conj().T
-    transition = dipole_transitions(stack, add_induced=add_induced)
-    rho, rhos = jax.lax.scan(
-        integrate, stack.eigenvectors @ stack.rho_0 @ stack.eigenvectors.conj().T, time
-    )
-
-    return (
-        stack.replace(rho_0=stack.eigenvectors.conj().T @ rho @ stack.eigenvectors),
-        rhos,
-    )
-
-
-## INDEPENDENT-PARTICLE CALCULATIONS
-def transition_energies(stack: Stack) -> jax.Array:
-    """Computes independent-particle transition energies associated with the TB-Hamiltonian of a stack.
-
-        - `stack`:
-
-    **Returns:**
-
-     array, the element `arr[i,j]` contains the transition energy from `i` to `j`
-    """
-    return jnp.abs(jnp.expand_dims(stack.energies, 1) - stack.energies)
-
-
-def wigner_weisskopf(stack: Stack, component: int = 0) -> jax.Array:
-    """Calculcates Wigner-Weisskopf transiton rates.
-
-        - `stack`:
-        - `component`: component of the dipolar transition to take $(0,1,2) \rightarrow (x,y,z)$`
-
-
-    **Returns:**
-
-     array, the element `arr[i,j]` contains the transition rate from `i` to `j`
-    """
-    charge = 1.602e-19
-    eps_0 = 8.85 * 1e-12
-    hbar = 1.0545718 * 1e-34
-    c = 3e8  # 137 (a.u.)
-    factor = 1.6e-29 * charge / (3 * jnp.pi * eps_0 * hbar**2 * c**3)
-    te = transition_energies(stack)
-    return (
-        (te * (te > stack.eps)) ** 3
-        * jnp.squeeze(transition_dipole_moments(stack)[:, :, component] ** 2)
-        * factor
-    )
-
-
-# TODO: this should be changed in accordance with the newly defined position operator
-def transition_dipole_moments(stack: Stack) -> jax.Array:
-    r"""Compute transition dipole moments for all states $i,j$ as $<i | \hat{r} | j>$.
-
-        - `stack`: stack object with N orbitals
-
-    **Returns:**
-
-     dipole moments as a complex $N \\times N \\times 3$ - matrix, where the last component is the direction of the dipole moment.
-    """
-    return jnp.einsum(
-        "li,lj,lk", stack.eigenvectors.conj(), stack.eigenvectors, stack.positions
-    ) * jnp.expand_dims(
-        jnp.ones_like(stack.eigenvectors) - jnp.eye(stack.eigenvectors.shape[0]), 2
-    )
-
-
-## INTERACTION
-def epi(stack: Stack, rho: jax.Array, omega: float, epsilon: float = None) -> float:
-    r"""Calculates the EPI (Energy-based plasmonicity index) of a mode at $\hbar\omega$ in the absorption spectrum of a structure.
-
-        - `stack`: stack object
-        - `rho`: density matrix
-        - `omega`: energy at which the system has been CW-illuminated ($\hbar\omega$ in eV)
-        - `epsilon`: small broadening parameter to ensure numerical stability, if `None`, stack.eps is chosen
-
-    **Returns:**
-
-    , a number between `0` (single-particle-like) and `1` (plasmonic).
-    """
-    epsilon = stack.eps if epsilon is None else epsilon
-    rho_without_diagonal = jnp.abs(rho - jnp.diag(jnp.diag(rho)))
-    rho_normalized = rho_without_diagonal / jnp.linalg.norm(rho_without_diagonal)
-    te = transition_energies(stack)
-    excitonic_transitions = (
-        rho_normalized / (te * (te > stack.eps) - omega + 1j * epsilon) ** 2
-    )
-    return 1 - jnp.sum(jnp.abs(excitonic_transitions * rho_normalized)) / (
-        jnp.linalg.norm(rho_normalized) * jnp.linalg.norm(excitonic_transitions)
-    )
-
-
-# TODO: get t-matrix in some way, degeneracy identification inconsistent with user-defined eps
 def rpa_polarizability_function(
     stack, tau, polarization, coulomb_strength, phi_ext=None, hungry=True
 ):
@@ -1771,73 +575,7 @@ def bare_susceptibility_function(stack, tau, hungry=True):
     return _susceptibility
 
 
-## POSTPROCESSING
-def indices(stack: Stack, orbital_id: str) -> jax.Array:
-    """Gets indices of a specific orbital.
-
-        Can be used to calculate, e.g. positions and energies corresponding to that orbital in the stack.
-
-        - `stack`: stack object
-        - `orbital_id`: orbital identifier as contained in :class:`granad.numerics.Stack.unique_ids`
-
-    **Returns:**
-
-     array corresponding to the indices of the orbitals, such that e.g. `stack.energies[indices]` gives the energies associated with the orbitals.
-    """
-    return jnp.nonzero(stack.ids == stack.unique_ids.index(orbital_id))[0]
-
-
-# TODO: this should be changed in accordance with the newly defined position operator
-def induced_dipole_moment(stack: Stack, rhos_diag: jax.Array) -> jax.Array:
-    """
-        Calculates the induced dipole moment for a collection of density matrices.
-
-        - `stack`: stack object
-        - `rhos_diag`: $N \\times m$ time-dependent site occupation matrix, indexed by rhos_diag[timestep,site_number]
-
-    **Returns:**
-
-    , $N \\times 3$ matrix containing the induced dipole moment $(p_x, p_y, p_z)$ at $N$ times
-    """
-    return (
-        (jnp.diag(to_site_basis(stack, stack.rho_stat)) - rhos_diag).real
-        @ stack.positions
-        * stack.electrons
-    )
-
-
-def induced_field(
-    stack: Stack, positions: jax.Array, density_matrix: Union[jax.Array, None] = None
-) -> jax.Array:
-    """Classical approximation to the induced (local) field in a stack.
-
-    - `stack`: a stack object
-    - `positions`: positions to evaluate the field on, must be of shape N x 3
-    :density_matrix: if given, compute the field corresponding to this density matrix. otherwise, use `stack.rho_0`.
-    """
-
-    # determine whether to use argument or the state of the stack
-    density_matrix = density_matrix if density_matrix is not None else stack.rho_0
-
-    # distance vector array from field sources to positions to evaluate field on
-    vec_r = stack.positions[:, None] - positions
-
-    # scalar distances
-    denominator = jnp.linalg.norm(vec_r, axis=2) ** 3
-
-    # normalize distance vector array
-    point_charge = jnp.nan_to_num(
-        vec_r / denominator[:, :, None], posinf=0.0, neginf=0.0
-    )
-
-    # compute charge via occupations in site basis
-    charge = stack.electrons * jnp.diag(to_site_basis(stack, density_matrix)).real
-
-    # induced field is a sum of point charges, i.e. \vec{r} / r^3
-    e_field = 14.39 * jnp.sum(point_charge * charge[:, None, None], axis=0)
-    return e_field
-
-
+# TODO: think of smarter of doing this
 def to_site_basis(stack: Stack, matrix: jax.Array) -> jax.Array:
     """Transforms an arbitrary matrix from energy to site basis.
